backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
alembic==1.16.4
psycopg2-binary==2.9.9; python_version < "3.10"
psycopg2-binary==2.9.10; python_version >= "3.10"
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
structlog==25.4.0
redis==6.2.0
celery==5.5.3
qdrant-client==1.15.1
minio==7.2.18
httpx==0.28.1
dateparser==1.2.0
langgraph>=0.2.0,<0.7.0
langchain-core>=0.3.0,<0.4.0
langgraph-checkpoint-postgres>=1.0.0,<2.0.0
cryptography==45.0.6
email-validator==2.2.0
pypdf==5.9.0
python-docx==1.2.0
ruff==0.12.8
pytest==8.4.1
pytest-cov==6.2.1
